+++ frontend/src/app/HeaderContent/index.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { useDispatch } from 'react-redux';
 
 import { Avatar, Menu, Dropdown } from 'antd';
+import Notifications from '@/components/Navigaton';
 
 import {
   AppstoreOutlined,
@@ -21,33 +22,7 @@ export default function HeaderContent() {
   const dispatch = useDispatch();
   const { SubMenu } = Menu;
   
-  const Notifications = () => {
-
-  const [notifications, setNotifications] = React.useState([
-    {id: 1, text: 'First notification'},
-    {id: 2, text: 'Second notification'},
-    {id: 3, text: 'Third notification'}  
-  ]);
-
-  const deleteNotification = (id) => {
-    const updatedNotifications = notifications.filter(n => n.id !== id);
-    setNotifications(updatedNotifications);
-  }
-
-  return (
-    <div className="notifications whiteBox shadow" style={{ minWidth: '200px' }}>
-      {notifications.map((n) => (
-        <div key={n.id} className="notification pad15">
-          {n.text}
-          <button onClick={() => deleteNotification(n.id)} className="info">
-            <DeleteOutlined />
-          </button>
-        </div>
-      ))}
-    </div>
-  );
-}
-
+  
 
   const profileDropdown = (
     <div className="profileDropdown whiteBox shadow" style={{ minWidth: '200px' }}>
